teste: implementa inicial de testes da aplicação.

+++ back-end/src/Store.App.Core.Application/Produto/Apagar/ApagarProdutoHandler.cs
@@ -1,5 +1,6 @@
 ﻿using MediatR;
 using Store.App.Core.Domain.Entitites;
+using Store.App.Core.Domain.Enum;
 using Store.App.Core.Domain.Repositories;
 using Store.App.Core.Domain.Repositories.Carro;
 
@@ -19,8 +20,28 @@ namespace Store.App.Core.Application.Produto.Apagar
         {
             ProdutoEntity produtoEntity = await _produtoRepository.Selecionar(x => x.Id == request.Id, cancellationToken, "Carro");
 
+            if(produtoEntity is null)
+            {
+                return new ApagarProdutoResponse
+                {
+                    StatusCode = System.Net.HttpStatusCode.NotFound,
+                    TextResponse = "Produto não encontrado."
+                };
+            }
+
             if(produtoEntity.Carro.Count > 0)
             {
+                bool produtoJaTeveCompraRealizada = produtoEntity.Carro.Any(x => x.Status == StatusCompraCarro.COMPRA_REALIZADA);
+
+                if (produtoJaTeveCompraRealizada)
+                {
+                    return new ApagarProdutoResponse
+                    {
+                        StatusCode = System.Net.HttpStatusCode.BadRequest,
+                        TextResponse = "Produto não pode ser excluído, pois já teve compra realizada."
+                    };
+                }
+
                 _carroRepository.RemoveRange(produtoEntity.Carro);
                 await _carroRepository.SaveChangesAsync(cancellationToken);
             }
@@ -28,7 +49,11 @@ namespace Store.App.Core.Application.Produto.Apagar
             _produtoRepository.Apagar(new ProdutoEntity { Id = request.Id });
             await _produtoRepository.SaveChangesAsync(cancellationToken);
 
-            return new ApagarProdutoResponse();
+            return new ApagarProdutoResponse
+            {
+                StatusCode = System.Net.HttpStatusCode.OK,
+                TextResponse = "Produto apagado com sucesso!"
+            };
         }
     }
 }
